UI tweaks, audio fix
