--- DEVTEST/game/G_engine.txt
@@ -1,16 +1,0 @@
-TTY Adventure Game Parser
-
-
-This is a simple text parser for text adventure games, designed to be compatible with 
-TTY/TDD devices. It is made with the Ultratec Superprint 4425 in mind, but should theoretically
-work with any TDD, or even over a terminal.
-
-The main restrictions are:
-- limit of 24 characters per line
-- Reduced character set (both in and out)
-- File size (due to running on a Pi Pico)
-- low transfer speeds (45.5 baud!!)
-
-
-
-
