--- DEVTEST/game/scenes.json
@@ -1,27 +1,0 @@
-{ 
-    "start":{
-        "roomid":0,
-        "disp_text":"WELCOME TO MY TEST GAME\nTHERE ARE LIMITED\nOPTIONS AVAILABLE\nWE ARE TESTING BASICS\n\nYOU STAND AT A DOOR",
-        "acts":{
-            "OPEN":"dooropen",
-            "N":"block",
-            "S":"block",
-            "E":"block",
-            "W":"block"
-        }
-
-    },
-    "dooropen":{
-        "roomid":1,
-        "disp_text":"THE DOOR IS OPEN! YOU WIN!",
-        "acts":{},
-        "win":1
-    },
-    "block":{
-        "roomid":2,
-        "disp_text":"NO.",
-        "acts":{
-            "BACK":"start"
-        }
-    }
-}
