--- DEVTEST/game/script.txt
@@ -1,12 +1,0 @@
-BBBBBBBBBBBBBBBBBBBBBBB
-MB------------------- B
-BB-------------------BB
-BB-------------------BB
-BB-------------------BB
-BB-------------------BB
-BBBBBBBBBBBBBBBBBBBBBBB
-BBBB++++++++++++++BBBBB
-BBBB++BBBB++++++++BBBBB
-BBBB++BBBB++++++++BBBBB
-BBBB++BBBB++++++++BBBBB
-BBBB++++++++++++++BBBBB
