+++ wav/output/testchars.txt
@@ -1,0 +1,6 @@
+"TESTOUT.WAV" should output the following:
+
+ABCDEFGHIJKLM
+NOPQRSTUVWXYZ
+1 2 3 4 5 6 7 8 9 0
+()-=+!$/?.,';":
